requirements.txt
flask
tensorflow
joblib
gunicorn
scikit-learn
Pillow
numpy